requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
greenlet>=3.0.0

# Redis
redis>=5.0.0
aioredis>=2.0.0

# Data validation and serialization
pydantic>=2.5.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP client for testing
httpx>=0.25.0

# OpenAI API client
openai>=1.3.0

# Environment variables
python-dotenv>=1.0.0

# AI agents framework
openai-agents>=0.1.0

# Async support
asyncio

# Standard library dependencies (included with Python)
# - dataclasses
# - enum
# - typing
# - threading
# - json
